Moved registration intent to apps package
